rename: 예매 내역 페이지 이름 변경
