+++ app/src/main/java/woowacourse/movie/feature/NavigationFragmentType.kt
@@ -3,19 +3,19 @@ package woowacourse.movie.feature
 import androidx.fragment.app.Fragment
 import woowacourse.movie.R
 import woowacourse.movie.feature.home.movie.MovieHomeFragment
-import woowacourse.movie.feature.reservationlist.ReservationListFragment
+import woowacourse.movie.feature.history.ReservationHistoryFragment
 import woowacourse.movie.feature.setting.SettingFragment
 import java.lang.IllegalArgumentException
 
 enum class NavigationFragmentType(val tag: String) {
-    RESERVATION_LIST("reservationListFragment"),
+    RESERVATION_HISTORY("reservationHistoryFragment"),
     MOVIE_HOME("movieHomeFragment"),
     SETTING("settingFragment"),
     ;
 
     fun newInstance(): Fragment {
         return when (this) {
-            RESERVATION_LIST -> ReservationListFragment()
+            RESERVATION_HISTORY -> ReservationHistoryFragment()
             MOVIE_HOME -> MovieHomeFragment()
             SETTING -> SettingFragment()
         }
@@ -26,7 +26,7 @@ enum class NavigationFragmentType(val tag: String) {
 
         fun from(menuItemId: Int): NavigationFragmentType {
             return when (menuItemId) {
-                R.id.reservation_list_item -> RESERVATION_LIST
+                R.id.reservation_history_item -> RESERVATION_HISTORY
                 R.id.home_item -> MOVIE_HOME
                 R.id.setting_item -> SETTING
                 else -> throw IllegalArgumentException(INVALID_MENU_ITEM_ID_MESSAGE)
