+++ app/src/main/res/menu/bottom_navigation_items.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/reservation_list_item"
+        android:id="@+id/reservation_history_item"
         android:icon="@drawable/ic_list"
         android:title="예매 내역"/>
